Solution to Assignment
Solution:
Please refer to the attached Program_01 to Program_06 and its respective Screenshot_Program_01 to Screenshot_Program_06.

Problem Statements:
1) Write a java code with the class named ‘acad’ and a method ‘main’. Hard Code the program
with two integers and print the sum of those two.

2) Rewrite the above code, where, inputs are provided by the user at runtime and the output is
printed.

3) Write a program with method name sum() that accepts two parameters from user and print
the sum of two numbers. Output format should be as:

First number is:
Second number is:
Sum is:

4) Write a program to accepts two numbers from stdin and find all the odd as well as even
numbers present in between them.

5) Joe is scared to go to school. When her dad asked the reason, joe said she is unable to
complete the task given by her teacher. The task was to find the “first 10 multiples” of the
number entered from stdin . Eg:
Input: 3
O/p:
3 x 1 = 3

3 x 2 = 6